lxd/project/limits: Move content from lxd/project
