+++ lxd/project/limits/state.go
@@ -1,4 +1,4 @@
-package project
+package limits
 
 import (
 	"context"
